+++ Obsidian/Entities/Player.cs
@@ -2,7 +2,6 @@
 // https://wiki.vg/Map_Format
 using Microsoft.CodeAnalysis;
 using Microsoft.Extensions.Logging;
-using Obsidian.API._Types;
 using Obsidian.API.Events;
 using Obsidian.API.Utilities;
 using Obsidian.Nbt;
@@ -13,11 +12,9 @@ using Obsidian.Net.Packets.Play.Clientbound;
 using Obsidian.Net.Scoreboard;
 using Obsidian.Registries;
 using Obsidian.WorldData;
-using System.Buffers;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
-using System.Linq;
 using System.Net;
 
 namespace Obsidian.Entities;
@@ -32,7 +29,7 @@ public sealed partial class Player : Living, IPlayer
     /// <summary>
     /// Used to log actions caused by the client.
     /// </summary>
-    protected ILogger Logger { get; private set; }
+    private ILogger Logger => this.client.Logger;
 
     internal HashSet<IPlayer> visiblePlayers = [];
 
@@ -169,29 +166,6 @@ public sealed partial class Player : Living, IPlayer
     public ItemStack? GetHeldItem() => Inventory.GetItem(inventorySlot);
     public ItemStack? GetOffHandItem() => Inventory.GetItem(45);
 
-    public async Task LoadPermsAsync()
-    {
-        // Load a JSON file that contains all permissions
-        var file = new FileInfo(Path.Combine(Server.PermissionPath, $"{Uuid}.json"));
-
-        if (file.Exists)
-        {
-            await using var fs = file.OpenRead();
-            if (await fs.FromJsonAsync<Permission>() is Permission permission)
-                PlayerPermissions = permission;
-        }
-    }
-
-    public async Task SavePermsAsync()
-    {
-        // Save permissions to JSON file
-        var file = new FileInfo(Path.Combine(Server.PermissionPath, $"{Uuid}.json"));
-
-        await using var fs = file.Open(FileMode.OpenOrCreate, FileAccess.ReadWrite);
-
-        await PlayerPermissions.ToJsonAsync(fs);
-    }
-
     public async Task DisplayScoreboardAsync(IScoreboard scoreboard, ScoreboardPosition position)//TODO implement new features
     {
         var actualBoard = (Scoreboard)scoreboard;
@@ -305,19 +279,19 @@ public sealed partial class Player : Living, IPlayer
         // reload world stuff and send rest of the info
         await UpdateChunksAsync(true, 2);
 
-        await client.SendInfoAsync();
+        await SendInitialInfoAsync();
 
         var (chunkX, chunkZ) = Position.ToChunkCoord();
         await client.QueuePacketAsync(new SetCenterChunkPacket(chunkX, chunkZ));
     }
 
-    public Task SendMessageAsync(ChatMessage message, Guid sender, SecureMessageSignature messageSignature) =>
+    public ValueTask SendMessageAsync(ChatMessage message, Guid sender, SecureMessageSignature messageSignature) =>
         throw new NotImplementedException();
 
-    public Task SendMessageAsync(ChatMessage message) =>
+    public ValueTask SendMessageAsync(ChatMessage message) =>
         client.QueuePacketAsync(new SystemChatMessagePacket(message, false));
 
-    public Task SetActionBarTextAsync(ChatMessage message) =>
+    public ValueTask SetActionBarTextAsync(ChatMessage message) =>
         client.QueuePacketAsync(new SystemChatMessagePacket(message, true));
 
     public async Task SendSoundAsync(ISoundEffect soundEffect)
@@ -549,7 +523,6 @@ public sealed partial class Player : Living, IPlayer
         float offsetY, float offsetZ, float extra = 0) =>
         SpawnParticleAsync(particle, new VectorF(x, y, z), count, offsetX, offsetY, offsetZ, extra);
 
-
     public async Task SpawnParticleAsync(ParticleType particle, VectorF pos, int count, float extra = 0) =>
         await client.QueuePacketAsync(new ParticlePacket
         {
@@ -599,187 +572,6 @@ public sealed partial class Player : Living, IPlayer
             Offset = new VectorF(offsetX, offsetY, offsetZ),
             MaxSpeed = extra
         });
-
-    public async Task SaveAsync()
-    {
-        var playerDataFile = new FileInfo(GetPlayerDataPath());
-        var persistentDataFile = new FileInfo(PersistentDataFile);
-
-        if (playerDataFile.Exists)
-        {
-            playerDataFile.CopyTo(GetPlayerDataPath(true), true);
-            playerDataFile.Delete();
-        }
-
-        if (persistentDataFile.Exists)
-        {
-            persistentDataFile.CopyTo(PersistentDataBackupFile, true);
-            persistentDataFile.Delete();
-        }
-
-        await using var persistentDataStream = persistentDataFile.Create();
-        await using var persistentDataWriter = new NbtWriter(persistentDataStream, NbtCompression.GZip, "");
-
-        persistentDataWriter.WriteString("worldName", world.ParentWorldName ?? world.Name);
-        //TODO make sure to save inventory in the right location if has using global data set to true
-
-        persistentDataWriter.EndCompound();
-        await persistentDataWriter.TryFinishAsync();
-
-        await using var playerFileStream = playerDataFile.Create();
-        await using var writer = new NbtWriter(playerFileStream, NbtCompression.GZip, "");
-
-        writer.WriteInt("DataVersion", 3337);
-        writer.WriteInt("playerGameType", (int)Gamemode);
-        writer.WriteInt("previousPlayerGameType", (int)Gamemode);
-        writer.WriteInt("Score", 0);
-        writer.WriteInt("SelectedItemSlot", inventorySlot);
-        writer.WriteInt("foodLevel", FoodLevel);
-        writer.WriteInt("foodTickTimer", FoodTickTimer);
-        writer.WriteInt("XpLevel", XpLevel);
-        writer.WriteInt("XpTotal", XpTotal);
-
-        writer.WriteFloat("Health", Health);
-
-        writer.WriteFloat("foodExhaustionLevel", FoodExhaustionLevel);
-        writer.WriteFloat("foodSaturationLevel", FoodSaturationLevel);
-
-        writer.WriteString("Dimension", world.DimensionName);
-
-        writer.WriteListStart("Pos", NbtTagType.Double, 3);
-
-        writer.WriteDouble(Position.X);
-        writer.WriteDouble(Position.Y);
-        writer.WriteDouble(Position.Z);
-
-        writer.EndList();
-
-        writer.WriteListStart("Rotation", NbtTagType.Float, 2);
-
-        writer.WriteFloat(Yaw);
-        writer.WriteFloat(Pitch);
-
-        writer.EndList();
-
-        WriteItems(writer);
-        WriteItems(writer, false);
-
-        writer.EndCompound();
-
-        await writer.TryFinishAsync();
-    }
-
-    public async Task LoadAsync(bool loadFromPersistentWorld = true)
-    {
-        // Read persistent data first
-        var persistentDataFile = new FileInfo(PersistentDataFile);
-
-        if (persistentDataFile.Exists)
-        {
-            await using var persistentDataStream = persistentDataFile.OpenRead();
-
-            var persistentDataReader = new NbtReader(persistentDataStream, NbtCompression.GZip);
-
-            //TODO use inventory if has using global data set to true
-            if (persistentDataReader.ReadNextTag() is NbtCompound persistentDataCompound)
-            {
-                var worldName = persistentDataCompound.GetString("worldName")!;
-
-                Logger?.LogInformation("persistent world: {worldName}", worldName);
-
-                if (loadFromPersistentWorld && this.world.WorldManager.TryGetWorld<World>(worldName, out var world))
-                {
-                    base.world = world;
-                    Logger?.LogInformation("Loading from persistent world: {worldName}", worldName);
-                }
-            }
-        }
-
-        // Then read player data
-        var playerDataFile = new FileInfo(GetPlayerDataPath());
-
-        await LoadPermsAsync();
-
-        if (!playerDataFile.Exists)
-        {
-            Position = world.LevelData.SpawnPosition;
-            return;
-        }
-
-        await using var playerFileStream = playerDataFile.OpenRead();
-
-        var reader = new NbtReader(playerFileStream, NbtCompression.GZip);
-
-        var compound = reader.ReadNextTag() as NbtCompound;
-        Debug.Assert(compound is not null); // TODO Handle invalid NBT
-
-        OnGround = compound.GetBool("OnGround");
-        Sleeping = compound.GetBool("Sleeping");
-        Air = compound.GetShort("Air");
-        AttackTime = compound.GetShort("AttackTime");
-        DeathTime = compound.GetShort("DeathTime");
-        Health = compound.GetFloat("Health");
-        HurtTime = compound.GetShort("HurtTime");
-        SleepTimer = compound.GetShort("SleepTimer");
-        FoodLevel = compound.GetInt("foodLevel");
-        FoodTickTimer = compound.GetInt("foodTickTimer");
-        Gamemode = (Gamemode)compound.GetInt("playerGameType");
-        XpLevel = compound.GetInt("XpLevel");
-        XpTotal = compound.GetInt("XpTotal");
-        FallDistance = compound.GetFloat("FallDistance");
-        FoodExhaustionLevel = compound.GetFloat("foodExhaustionLevel");
-        FoodSaturationLevel = compound.GetFloat("foodSaturationLevel");
-        XpP = compound.GetInt("XpP");
-
-        if (!Alive)
-            Health = 20f;//Player should never load data that has health at 0 
-
-        var dimensionName = compound.GetString("Dimension");
-        if (!string.IsNullOrWhiteSpace(dimensionName) && CodecRegistry.TryGetDimension(dimensionName, out var codec))
-        {
-            //TODO load into dimension ^ ^
-        }
-
-        compound.TryGetTag("Pos", out var posTag);
-        Position = (posTag as NbtList) switch
-        {
-            [NbtTag<double> a, NbtTag<double> b, NbtTag<double> c, ..] => new VectorF((float)a.Value, (float)b.Value, (float)c.Value),
-            _ => world.LevelData.SpawnPosition
-        };
-
-        if (compound.TryGetTag("Rotation", out var rotationTag))
-        {
-            if (rotationTag is NbtList and [NbtTag<float> yaw, NbtTag<float> pitch, ..])
-            {
-                Yaw = yaw.Value;
-                Pitch = pitch.Value;
-            }
-        }
-
-        if (compound.TryGetTag("Inventory", out var rawTag) && rawTag is NbtList inventory)
-        {
-            foreach (var rawItemTag in inventory)
-            {
-                if (rawItemTag.Type == NbtTagType.End)
-                    break;
-
-                if (rawItemTag is not NbtCompound itemCompound)
-                    continue;
-
-                var slot = itemCompound.GetByte("Slot");
-
-                var itemMetaBuilder = new ItemMetaBuilder()
-                    .WithDurability(itemCompound.GetInt("Damage"))
-                    .IsUnbreakable(itemCompound.GetBool("Unbreakable"));
-
-                var item = ItemsRegistry.GetSingleItem(itemCompound.GetString("id"), itemMetaBuilder.Build());
-                item.Count = itemCompound.GetByte("Count");
-                item.Slot = slot;
-
-                Inventory.SetItem(slot, item);
-            }
-        }
-    }
 
     public async Task<bool> GrantPermissionAsync(string permissionNode)
     {
@@ -901,77 +693,6 @@ public sealed partial class Player : Living, IPlayer
         await PickupNearbyItemsAsync();
     }
 
-    private async ValueTask TrySpawnPlayerAsync(VectorF position)
-    {
-        //TODO PROPER DISTANCE CALCULATION
-        var entityBroadcastDistance = this.world.Configuration.EntityBroadcastRangePercentage;
-
-        foreach (var player in world.GetPlayersInRange(position, entityBroadcastDistance))
-        {
-            if (player == this)
-                continue;
-
-            if (player.Alive && !visiblePlayers.Contains(player))
-            {
-                visiblePlayers.Add(player);
-
-                player.SpawnEntity();
-            }
-        }
-
-        if (visiblePlayers.Count == 0)
-            return;
-
-        var removed = ArrayPool<int>.Shared.Rent(visiblePlayers.Count);
-
-        var index = 0;
-        visiblePlayers.RemoveWhere(visiblePlayer =>
-        {
-            if (!visiblePlayer.IsInRange(this, entityBroadcastDistance))
-            {
-                removed[index++] = visiblePlayer.EntityId;
-                return true;
-            }
-            return false;
-        });
-
-        if (index > 0)
-            await client.QueuePacketAsync(new RemoveEntitiesPacket(removed.ToArray()));
-
-        ArrayPool<int>.Shared.Return(removed);
-    }
-
-    private async Task PickupNearbyItemsAsync(float distance = 1.5f)
-    {
-        foreach (var entity in world.GetNonPlayerEntitiesInRange(Position, distance))
-        {
-            if (entity is not ItemEntity item)
-                continue;
-
-            if (!item.CanPickup)
-                continue;
-
-            this.PacketBroadcaster.QueuePacketToWorld(this.World, new PickupItemPacket
-            {
-                CollectedEntityId = item.EntityId,
-                CollectorEntityId = EntityId,
-                PickupItemCount = item.Count
-            });
-
-            var slot = Inventory.AddItem(new ItemStack(item.Material, item.Count, item.ItemMeta));
-
-            client.SendPacket(new SetContainerSlotPacket
-            {
-                Slot = (short)slot,
-                WindowId = 0,
-                SlotData = Inventory.GetItem(slot)!,
-                StateId = Inventory.StateId++
-            });
-
-            await item.RemoveAsync();
-        }
-    }
-
     /// <summary>
     /// Updates client chunks. Only send <paramref name="distance"/> when sending initial chunks.
     /// </summary>
@@ -1030,45 +751,4 @@ public sealed partial class Player : Living, IPlayer
 
         return sentAll;
     }
-
-    private void WriteItems(NbtWriter writer, bool inventory = true)
-    {
-        var items = inventory ? Inventory.Select((item, slot) => (item, slot)) : EnderInventory.Select((item, slot) => (item, slot));
-
-        var nonNullItems = items.Where(x => x.item != null);
-
-        writer.WriteListStart(inventory ? "Inventory" : "EnderItems", NbtTagType.Compound, nonNullItems.Count());
-
-        foreach (var (item, slot) in nonNullItems)
-        {
-            writer.WriteCompoundStart();
-
-            writer.WriteByte("Count", (byte)item.Count);
-            writer.WriteByte("Slot", (byte)slot);
-
-            writer.WriteString("id", item.AsItem().UnlocalizedName);
-
-            writer.WriteCompoundStart("tag");
-
-            writer.WriteInt("Damage", item.ItemMeta.Durability);
-            writer.WriteBool("Unbreakable", item.ItemMeta.Unbreakable);
-
-            //TODO: item attributes
-
-            writer.EndCompound();
-            writer.EndCompound();
-        }
-
-        if (!nonNullItems.Any())
-            writer.Write(NbtTagType.End);
-
-        writer.EndList();
-    }
-
-    private Dictionary<Guid, List<InfoAction>> CompilePlayerInfo(params InfoAction[] actions) => new()
-    {
-        { Uuid, actions.ToList() }
-    };
-
-    private string GetPlayerDataPath(bool isOld = false) => Path.Join(world.PlayerDataPath, isOld ? $"{Uuid}.dat.old" : $"{Uuid}.dat");
 }
